fix binding data services to dynamic services in some cases
Dynamic services require Syskit to have forms of "horizontal
inheritance", due to the fact that at deployment we might "merge" two
specialized models into one.

This shows up as two models that are not related by Ruby's hierarchy
relation, but are still related.

Fix a bunch of issues at different levels on the model representation
related to this case, and write corresponding tests.

Make sure we can cross-bind services from one to the other.